Splitting projection functions to new projection_endpoint
