+++ biocentral_server/embeddings/__init__.py
@@ -1,9 +1,11 @@
 from .embedding_task import EmbeddingTask
 from .embed import compute_embeddings_and_save_to_db
 from .embeddings_endpoint import embeddings_service_route
+from .projection_endpoint import projection_route
 
 __all__ = [
     'EmbeddingTask',
     'embeddings_service_route',
+    'projection_route',
     'compute_embeddings_and_save_to_db'
 ]
